+++ src/physics/object.rs
@@ -5,11 +5,11 @@ use crate::vector2::Vector2;
 // TODO: FixedVec?
 #[derive(Default)]
 pub struct ObjectSoa {
-    pub positions: Vec<Vector2<f32>>,
-    pub velocities: Vec<Vector2<f32>>,
-    pub accelerations: Vec<Vector2<f32>>,
-    pub radii: Vec<f32>,
-    pub masses: Vec<f32>,
+    pub positions: Vec<Vector2<f64>>,
+    pub velocities: Vec<Vector2<f64>>,
+    pub accelerations: Vec<Vector2<f64>>,
+    pub radii: Vec<f64>,
+    pub masses: Vec<f64>,
     pub colors: Vec<Option<Color>>,
     pub is_planet: Vec<bool>,
     pub planet_count: usize,
@@ -65,29 +65,29 @@ impl ObjectSoa {
 }
 
 pub struct ObjectSoaRef<'a> {
-    pub positions: &'a [Vector2<f32>],
-    pub velocities: &'a [Vector2<f32>],
-    pub accelerations: &'a [Vector2<f32>],
-    pub radii: &'a [f32],
-    pub masses: &'a [f32],
+    pub positions: &'a [Vector2<f64>],
+    pub velocities: &'a [Vector2<f64>],
+    pub accelerations: &'a [Vector2<f64>],
+    pub radii: &'a [f64],
+    pub masses: &'a [f64],
     pub colors: &'a [Option<Color>],
     pub is_planet: &'a [bool],
 }
 
 #[derive(Copy, Clone)]
 pub struct ObjectPrototype {
-    pub position: Vector2<f32>,
-    pub velocity: Vector2<f32>,
-    pub acceleration: Vector2<f32>,
-    pub radius: f32,
-    pub mass: f32,
+    pub position: Vector2<f64>,
+    pub velocity: Vector2<f64>,
+    pub acceleration: Vector2<f64>,
+    pub radius: f64,
+    pub mass: f64,
     pub color: Option<Color>,
     pub is_planet: bool,
 }
 
 impl ObjectPrototype {
     #[must_use]
-    pub fn new(position: Vector2<f32>) -> Self {
+    pub fn new(position: Vector2<f64>) -> Self {
         Self {
             position,
             velocity: Vector2::new(0.0, 0.0),
@@ -100,7 +100,7 @@ impl ObjectPrototype {
     }
 
     #[must_use]
-    pub fn momentum(&self) -> Vector2<f32> {
+    pub fn momentum(&self) -> Vector2<f64> {
         self.velocity * self.mass
     }
 
@@ -112,6 +112,6 @@ impl ObjectPrototype {
 
 #[derive(Copy, Clone)]
 pub struct ObjectUpdate {
-    pub position: Vector2<f32>,
-    pub velocity: Vector2<f32>,
+    pub position: Vector2<f64>,
+    pub velocity: Vector2<f64>,
 }
